Move KSTypeReferenceSyntheticImpl into common-util.
(cherry picked from commit 610f42d307753c8b5bbeecd7b5f9fe80eef16048)
