+++ compiler-plugin/src/main/kotlin/com/google/devtools/ksp/symbol/impl/synthetic/KSConstructorSyntheticImpl.kt
@@ -20,6 +20,7 @@ package com.google.devtools.ksp.symbol.impl.synthetic
 import com.google.devtools.ksp.KSObjectCache
 import com.google.devtools.ksp.isPublic
 import com.google.devtools.ksp.processing.impl.KSNameImpl
+import com.google.devtools.ksp.processing.impl.KSTypeReferenceSyntheticImpl
 import com.google.devtools.ksp.processing.impl.ResolverImpl
 import com.google.devtools.ksp.symbol.*
